run/requirements.txt
flask==2.3.3
gunicorn==21.2.0
werkzeug==2.3.7
nbconvert==6.5.4
nbformat==5.7.0
papermill==2.4.0
mistune==0.8.4
ipykernel==6.29.5
requests==2.26.0
google-cloud-secret-manager==2.8.0
gitpython==3.1.24
pyyaml==6.0.2
lxml[html_clean]==4.9.3
python-dotenv==1.0.0